+++ pyspaces/libc.py
@@ -5,5 +5,5 @@
 from ctypes import *
 
 
-libc = CDLL("libc.so.6")
+libc = CDLL("libc.so.6", use_errno=True)
 """Import libc.so.6 as libc"""