+++ pyspaces/__init__.py
@@ -17,7 +17,7 @@ __license__ = "MIT"
 __maintainer__ = "Filipp Frizzy"
 __email__ = "filipp.s.frizzy@gmail.com"
 __status__ = "Development"
-__version__ = '1.2.3'
+__version__ = '1.2.4'
 
 __all__ = ["cloning", "process", "libc", "cli"]